final forms: json of form 204 data
final forms is a json that maps case ID to a tuple that is a list of creditors and a list of amounts. I did not match creditors to amounts due to noise in the data (sometimes — and unpredictably — they are off by one, creating misleading information), but it would be easy to match them if more spot-checking reveals higher accuracy than expected.